+++ client/src/app/modules/image-viewers/widgets/context-image/context-image.component.ts
@@ -338,6 +338,21 @@ export class ContextImageComponent extends BaseWidgetModel implements OnInit, On
             this.mdl.transform.scale.y = 1;
           }
 
+          if (contextData.colourRatioMin) {
+            this.mdl.colourRatioMin = contextData.colourRatioMin;
+          }
+
+          if (contextData.colourRatioMax) {
+            this.mdl.colourRatioMax = contextData.colourRatioMax;
+          }
+
+          this.mdl.imageBrightness = contextData.brightness;
+          this.mdl.removeTopSpecularArtifacts = contextData.removeTopSpecularArtifacts;
+          this.mdl.removeBottomSpecularArtifacts = contextData.removeBottomSpecularArtifacts;
+          this.mdl.rgbuChannels = contextData.rgbuChannels;
+          this.mdl.unselectedOpacity = contextData.unselectedOpacity;
+          this.mdl.unselectedGrayscale = contextData.unselectedGrayscale;
+
           this.mdl.imageName = contextData.contextImage;
           this.mdl.imageSmoothing = contextData.contextImageSmoothing.length > 0;
 
@@ -728,7 +743,7 @@ export class ContextImageComponent extends BaseWidgetModel implements OnInit, On
       .pipe(
         switchMap((data: ContextImageModelLoadedData) => {
           if (data.scanModels.size > 0) {
-            this.scanId = data.scanModels.keys().next().value;
+            this.scanId = data.scanModels.keys().next().value!;
           }
 
           this.mdl.setData(data);